Add customization for variable pitch font
The doom-nebula fonts use the doom variable pitch font in several faces.
I like to use a large, thin font for headings to contrast with the fixed
width font. With this change, users can now specify the weight as well
as a scale factor for height (size) when these faces are used.